pyproject.toml
[tool.poetry]
name = "file-manager-in-django"
version = "0.1.0"
description = ""
authors = ["Mccartheney Mendes <mccartheney@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
